+++ The HELIOS System/Assets/Dialogue/Level 1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 817021987ad0f674daa92ba1e6489cba
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
