+++ The HELIOS System/Assets/Dialogue/Level 1/DialogueAsset1.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b5cb8e3199cf35b4b9e411fe37902f6e, type: 3}
+  m_Name: DialogueAsset1
+  m_EditorClassIdentifier: 
+  dialogue:
+  - 'Welcome to Earth, I am your helper. My job is to introduce you to Earth''s ecosystems. '
+  - 'Please, take a moment to look at the ecosystem. '
+  - ...
+  - 'It''s looking a little empty right now, isn''t it? '
+  - 'How about you plant a tree or something. '
