+++ The HELIOS System/Assets/Dialogue/Level 1/DialogueAsset1.asset.meta
