+++ The HELIOS System/Assets/Dialogue/Level 1/DialogueAsset2.asset
@@ -10,10 +10,11 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: b5cb8e3199cf35b4b9e411fe37902f6e, type: 3}
-  m_Name: DialogueAsset1
+  m_Name: DialogueAsset2
   m_EditorClassIdentifier: 
   dialogue:
-  - Welcome to the game
-  - You are an AI, welcome to Earth
-  - Welcome v2
-  - Welcome version 3
+  - Congratulations on planting the tree! This is the first step to learning about
+    an ecosystem
+  - 'Trees and other plants harness energy from the sun. This energy is passed to
+    the animals that eat the plants. '
+  - So by planting the tree, you are setting the foundation of an ecosystem!
