+++ The HELIOS System/Assets/Dialogue/Level 1/EndingDialogueAsset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa5c96810dd6fb446aaef5c9018a604b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
