+++ The HELIOS System/Assets/Scripts/StageManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e6e89d87e6f704a4caebd03e67dbc509
